Move autovacuum tests in the same class
