+++ test_queue_job/tests/test_job.py
@@ -25,14 +25,11 @@ from odoo.addons.queue_job.job import (
     identity_exact,
 )
 
+from .common import JobCommonCase
 
-class TestJobsOnTestingMethod(common.TransactionCase):
+
+class TestJobsOnTestingMethod(JobCommonCase):
     """ Test Job """
-
-    def setUp(self):
-        super(TestJobsOnTestingMethod, self).setUp()
-        self.queue_job = self.env["queue.job"]
-        self.method = self.env["test.queue.job"].testing_method
 
     def test_new_job(self):
         """
@@ -314,7 +311,7 @@ class TestJobsOnTestingMethod(common.TransactionCase):
         self.assertEqual(job1.identity_key, expected_key)
 
 
-class TestJobs(common.TransactionCase):
+class TestJobs(JobCommonCase):
     """ Test jobs on other methods or with different job configuration """
 
     def test_description(self):
@@ -420,23 +417,7 @@ class TestJobs(common.TransactionCase):
         self.assertEquals({"mutable_kwarg": {"a": 1}}, job_instance.kwargs)
 
 
-class TestJobModel(common.TransactionCase):
-    def setUp(self):
-        super(TestJobModel, self).setUp()
-        self.queue_job = self.env["queue.job"]
-        self.user = self.env["res.users"]
-        self.method = self.env["test.queue.job"].testing_method
-        self.env["queue.job.function"]._register_job(
-            self.env["test.queue.job"], self.method
-        )
-
-    def _create_job(self):
-        test_job = Job(self.method)
-        test_job.store()
-        stored = Job.db_record_from_uuid(self.env, test_job.uuid)
-        self.assertEqual(len(stored), 1)
-        return stored
-
+class TestJobModel(JobCommonCase):
     def test_job_change_state(self):
         stored = self._create_job()
         stored._change_job_state(DONE, result="test")
@@ -488,39 +469,6 @@ class TestJobModel(common.TransactionCase):
         followers = stored.message_follower_ids.mapped("partner_id")
         self.assertFalse(inactiveusr.partner_id in followers)
         self.assertFalse({u.partner_id for u in group.users} - set(followers))
-
-    def test_autovacuum(self):
-        # test default removal interval
-        stored = self._create_job()
-        date_done = datetime.now() - timedelta(days=29)
-        stored.write({"date_done": date_done})
-        self.env["queue.job"].autovacuum()
-        self.assertEqual(len(self.env["queue.job"].search([])), 1)
-
-        date_done = datetime.now() - timedelta(days=31)
-        stored.write({"date_done": date_done})
-        self.env["queue.job"].autovacuum()
-        self.assertEqual(len(self.env["queue.job"].search([])), 0)
-
-    def test_autovacuum_multi_channel(self):
-        root_channel = self.env.ref("queue_job.channel_root")
-        channel_60days = self.env["queue.job.channel"].create(
-            {"name": "60days", "removal_interval": 60, "parent_id": root_channel.id}
-        )
-        date_done = datetime.now() - timedelta(days=31)
-        job_root = self._create_job()
-        job_root.write({"date_done": date_done})
-        job_60days = self._create_job()
-        job_60days.write({"channel": channel_60days.name, "date_done": date_done})
-
-        self.assertEqual(len(self.env["queue.job"].search([])), 2)
-        self.env["queue.job"].autovacuum()
-        self.assertEqual(len(self.env["queue.job"].search([])), 1)
-
-        date_done = datetime.now() - timedelta(days=61)
-        job_60days.write({"date_done": date_done})
-        self.env["queue.job"].autovacuum()
-        self.assertEqual(len(self.env["queue.job"].search([])), 0)
 
     def test_wizard_requeue(self):
         stored = self._create_job()
